WIP start on mesh wrapper [skip ci]
